+++ framework/db-local/src/test/java/is/codion/framework/db/local/TestDomain.java
@@ -192,11 +192,14 @@ public final class TestDomain extends DomainModel {
 										Employee.SALARY.define()
 														.column()
 														.caption(Employee.SALARY.name())
-														.nullable(false).valueRange(1000, 10000).maximumFractionDigits(2),
+														.nullable(false)
+														.valueRange(1000, 10000)
+														.maximumFractionDigits(2),
 										Employee.COMMISSION.define()
 														.column()
 														.caption(Employee.COMMISSION.name())
-														.valueRange(100, 2000).maximumFractionDigits(2),
+														.valueRange(100, 2000)
+														.maximumFractionDigits(2),
 										Employee.MGR.define()
 														.column(),
 										Employee.MGR_FK.define()