+++ framework/json-db/src/test/java/is/codion/framework/json/TestDomain.java
@@ -145,10 +145,13 @@ public final class TestDomain extends DomainModel {
 														.searchable(true),
 										Employee.SALARY.define()
 														.column()
-														.nullable(false).valueRange(1000, 10000).maximumFractionDigits(2),
+														.nullable(false)
+														.valueRange(1000, 10000)
+														.maximumFractionDigits(2),
 										Employee.COMMISSION.define()
 														.column()
-														.valueRange(100, 2000).maximumFractionDigits(2),
+														.valueRange(100, 2000)
+														.maximumFractionDigits(2),
 										Employee.MGR.define()
 														.column(),
 										Employee.MGR_FK.define()